Implement search controller which uses OTI
This search controller uses OpenTree Index (https://github.com/chinchliff/oti/) which
requires that neo4j be running and that the NexSON files have been indexed.

This controller also requires that the host and port of the neo4j instance be specified
in private/config. An example is given in private/config.example .